+++ publish/app/Http/Requests/Auth/Register/Request.php
@@ -19,6 +19,23 @@ class Request extends FormRequest
     }
 
     /**
+     *  Get custom messages for validator errors.
+     *
+     *  @return array
+     */
+    #[\Override]
+    public function messages()
+    {
+        if (config('enraiged.auth.allow_secondary_credential') === true) {
+            return [
+                'username.different' => 'The primary and secondary email addresses must be different.',
+            ];
+        }
+
+        return [];
+    }
+
+    /**
      *  Get the validation rules that apply to the request.
      *
      *  @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
@@ -56,7 +73,7 @@ class Request extends FormRequest
      *  @param  string  $table
      *  @return string
      */
-    private function transformEmailRule(string $rule, string $table): ?string
+    private function transformEmailRule(string $rule, string $table): string
     {
         if (config('enraiged.auth.allow_secondary_credential') === true) {
             return "{$rule}|unique:{$table},username";