+++ publish/resources/views/mail/auth/verify-secondary.blade.php
@@ -1,10 +1,10 @@
 <x-mail::message>
 # {{ __('Hello :name', ['name' => $user->name]) }}
 
-{{ __('Please click the button below to verify your email address.') }}
+{{ __('Please click the button below to verify your secondary email address.') }}
 
 <x-mail::button url="{{ $url }}">
-{{ __('Verify Email Address') }}
+{{ __('Verify Secondary Email Address') }}
 </x-mail::button>
 
 {{ __('If you did not create an account, no further action is required.') }}