+++ src/UserServiceProvider.php
@@ -4,6 +4,7 @@ namespace Enraiged;
 
 use Enraiged\Users\Observers\UserObserver;
 use Illuminate\Database\Eloquent\Relations\Relation;
+use Illuminate\Support\Facades\Event;
 use Illuminate\Support\ServiceProvider;
 
 class UserServiceProvider extends ServiceProvider
@@ -18,6 +19,21 @@ class UserServiceProvider extends ServiceProvider
         $model = config('auth.providers.users.model');
 
         $model::observe(UserObserver::class);
+
+        Event::listen(
+            \Illuminate\Auth\Events\Failed::class,
+            \Enraiged\Users\Listeners\FailedListener::class,
+        );
+
+        Event::listen(
+            \Illuminate\Auth\Events\Registered::class,
+            \Enraiged\Users\Listeners\RegisteredListener::class,
+        );
+
+        Event::listen(
+            \Illuminate\Auth\Events\Verified::class,
+            \Enraiged\Users\Listeners\VerifiedListener::class,
+        );
 
         Relation::morphMap([
             'user' => $model,