+++ src/Users/Listeners/RegisteredListener.php
@@ -13,7 +13,12 @@ class RegisteredListener
      */
     public function handle(Registered $event): void
     {
-        //  we will trigger the notification here if the system *does not* require email verification
+        //  we will trigger a secondary email verification notification here, if applicable
+        if (config('enraiged.auth.must_verify_secondary') === true && $event->user->usernameIsEmailAddress) {
+            $event->user->sendSecondaryVerificationNotification();
+        }
+
+        //  we will trigger a welcome notification here if the system *does not* require email verification
         if (config('enraiged.auth.must_verify_email') !== true) {
             $event->user->sendWelcomeNotification();
         }