+++ src/Users/Listeners/VerifiedListener.php
@@ -13,8 +13,8 @@ class VerifiedListener
      */
     public function handle(Verified $event): void
     {
-        //  we will trigger the notification here if the system *does* require email verification
-        if (config('enraiged.auth.must_verify_email') === true) {
+        //  we will trigger a welcome notification here if the system *does* require email verification
+        if (config('enraiged.auth.must_verify_email') === true && $event->user->wasRecentlySecondaryVerified) {
             $event->user->sendWelcomeNotification();
         }
     }