+++ src/Users/Models/Traits/MustVerifyEmail.php
@@ -3,10 +3,22 @@
 namespace Enraiged\Users\Models\Traits;
 
 use Enraiged\Users\Notifications\VerifyEmailNotification;
+use Illuminate\Contracts\Auth\MustVerifyEmail as VerificationContract;
 
 trait MustVerifyEmail
 {
     use \Illuminate\Auth\MustVerifyEmail;
+
+    /**
+     *  Determine whether this user must verify their email address.
+     *
+     *  @return bool
+     */
+    public function getMustVerifyEmailAttribute(): bool
+    {
+        return $this instanceof VerificationContract
+            && ! $this->hasVerifiedEmail();
+    }
 
     /**
      *  Send the email verification notification.