+++ src/Users/Models/VerifiedUser.php
@@ -7,14 +7,13 @@ use Illuminate\Contracts\Auth\MustVerifyEmail;
 
 class VerifiedUser extends User implements MustVerifyEmail
 {
-    use Traits\MustVerifyEmail;
-
     /**
      *  Create a new factory instance for the model.
      *
      *  @return \Illuminate\Database\Eloquent\Factories\Factory
      *  @static
      */
+    #[\Override]
     protected static function newFactory()
     {
         return new VerifiedUserFactory;