+++ src/Users/Notifications/ValidSecondaryLoginFailureNotification.php
@@ -7,7 +7,7 @@ use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Messages\MailMessage;
 
-class VerifyEmailNotification extends Notification implements ShouldQueue
+class ValidSecondaryLoginFailureNotification extends Notification implements ShouldQueue
 {
     use Queueable;
 
@@ -19,20 +19,18 @@ class VerifyEmailNotification extends Notification implements ShouldQueue
      */
     public function toMail($notifiable)
     {
-        $verificationUrl = $this->verificationUrl($notifiable);
-
         $message = (new MailMessage)
-            ->subject(__('Verify Email Address'));
+            ->subject(__('Secondary Login Failed'));
 
         if (config('enraiged.app.mail_markdown') === true) {
             return $message
-                ->markdown('mail.auth.verify-email', ['url' => $verificationUrl, 'user' => $notifiable]);
+                ->markdown('mail.auth.valid-secondary-failure', ['user' => $notifiable]);
         }
 
         return $message
             ->greeting(__('Hello :name', ['name' => $notifiable->name]))
-            ->line(__('Please click the button below to verify your email address.'))
-            ->action(__('Verify Email Address'), $verificationUrl)
-            ->line(__('If you did not create an account, no further action is required.'));
+            ->line('A recent attempt to log in with your secondary email address has failed.')
+            ->line('This email address must be validated before a log in will succeed.')
+            ->line('Please log in with your primary email address and navigate to your profile to send a new verification email.');
     }
 }