+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Release Notes
 
-2025-03-23 : 0.4.0.0-BETA  
+2025-03-26 : 0.4.0-beta.2 (prerelease)  
+Update: Added verification requirement for secondary email address
+
+- Added 'enraiged.auth.must_verify_secondary' config option
+- Added 'enraiged.auth.reject_unverified_secondary' config option
+- Added mail.auth templates for secondary address notifications
+- Removed accidental duplicate views from the publish directory
+- Updated language files as necessary
+
+---
+
+2025-03-23 : 0.4.0-beta.1 (prerelease)  
 Update: Completed assembling initial enraiged-core library